requirements.txt
Flask
Pillow
favicon
feedparser
pyyaml
tldextract
waitress
